fix photogallery page tests
